Consolidate linters into single binary
