+++ Makefile
@@ -311,9 +311,7 @@ contracts/test/prover/proofs/%.json: $(arbitrator_cases)/%.wasm $(arbitrator_pro
 # strategic rules to minimize dependency building
 
 .make/lint: $(DEP_PREDICATE) build-node-deps $(ORDER_ONLY_PREDICATE) .make
-	go run ./linter/koanf ./...
-	go run ./linter/pointercheck ./...
-	go run ./linter/rightshift ./...
+	go run ./linters ./...
 	golangci-lint run --fix
 	yarn --cwd contracts solhint
 	@touch $@
