+++ linters/koanf/handlers.go
@@ -1,4 +1,4 @@
-package main
+package koanf
 
 import (
 	"fmt"
