+++ linters/koanf/koanf_test.go
@@ -1,4 +1,4 @@
-package main
+package koanf
 
 import (
 	"errors"
