+++ linters/pointercheck/pointercheck_test.go
@@ -1,4 +1,4 @@
-package main
+package pointercheck
 
 import (
 	"os"
