+++ linters/rightshift/rightshift_test.go
@@ -1,4 +1,4 @@
-package main
+package rightshift
 
 import (
 	"os"
