+++ linters/structinit/structinit_test.go
@@ -1,4 +1,4 @@
-package main
+package structinit
 
 import (
 	"os"
