+++ linters/testdata/src/pointercheck/pointercheck.go
