+++ linters/testdata/src/rightshift/rightshift.go
