+++ linters/testdata/src/structinit/a/a.go
